+++ packages/typespec-go/CHANGELOG.md
@@ -10,8 +10,11 @@
 ### Features Added
 
 * Added switch `go-generate` to invoke post-generation scripts.
-  * The value is an output-relative path to a `.go` file that will can be used with `go generate`.
+  * The value is an output-relative path to a `.go` file containing `//go:generate` directives.
   * If Go tools are not on the path, and `go-generate` was specified, then an error is produced.
+
+### Other Changes
+
 * When Go tools are found on the path, the following steps happen after successfully generating code and any `go-generate` script is invoked.
   * Execute `gofmt -w .` followed by `go mod tidy` in the output directory.
   * If Go tools are not found, the above steps are skipped and a warning is displayed.